Code/requirements.txt
pytorch>=1.4.0
torchvision>=0.6.1
requests>=2.23.0
numpy>=1.18.5
pandas>=1.0.3
click>=7.1.2
click-config-file>=0.6.0
toml>=0.10.2
tqdm>=4.46.1
matplotlib>=3.2.2
pillow>=7.1.2
scikit-learn>=0.23.1
scikit-image>=0.16.2
scipy>=1.4.1
h5py>=2.10.0
scikit-video>=1.1.11
setuptools>=47.3.1
moviepy>=1.0.1
colorama>=0.4.4
